update mg27x branch based on master branch

+++ bin/MadGraph5_aMCatNLO/runcmsgrid_LO.sh
@@ -41,6 +41,9 @@ if [ "$use_gridpack_env" = true ]
     cd ${cmssw_version}/src
     eval `scramv1 runtime -sh`
 fi
+
+if [ ! -z "${PYTHON27PATH}" ] ; then export PYTHONPATH=${PYTHON27PATH} ; fi 
+
 cd $LHEWORKDIR/process
 
 #make sure lhapdf points to local cmssw installation area
@@ -110,10 +113,8 @@ done
 ls -lrt events*.lhe.gz
 if [  $run_counter -gt "1" ]; then
     echo "Merging files and deleting unmerged ones"
-    cp /cvmfs/cms.cern.ch/phys_generator/gridpacks/lhe_merger/merge.pl ./
-    chmod 755 merge.pl
-    # ./madevent/bin/internal/merge.pl events*.lhe.gz events.lhe.gz banner.txt
-    ./merge.pl events*.lhe.gz events.lhe.gz banner.txt
+    # use version in genproduction, not cvmfs, nor mg5_amc@nlo default 
+    perl $LHEWORKDIR/merge.pl events*.lhe.gz events.lhe.gz banner.txt
     rm events_*.lhe.gz banner.txt;
 else
     mv events_${run_counter}.lhe.gz events.lhe.gz
@@ -166,6 +167,13 @@ if [ -f ./madspin_card.dat ] ;then
     fi
 fi
 
+# Test if time_of_flight is set to a positive floating point value
+has_time_of_flight=$(egrep "^\s*\+?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?\s*=\s*time_of_flight" ./madevent/Cards/run_card.dat)
+if [ ! -z "$has_time_of_flight" ] ; then
+    ./madevent/bin/madevent add_time_of_flight events.lhe.gz
+fi
+    
+
 cd $LHEWORKDIR
 
 runlabel=GridRun_PostProc_${rnum}
@@ -177,17 +185,23 @@ if [ "$domadspin" -gt "0" ] ; then
 fi
 mv process/$event_file process/madevent/Events/${runlabel}/events.lhe
 
-# Add scale and PDF weights using systematics module
-#
+# add scale and PDF weights using systematics module
 pushd process/madevent
 pdfsets="PDF_SETS_REPLACE"
 scalevars="--mur=1,2,0.5 --muf=1,2,0.5 --together=muf,mur,dyn --dyn=-1,1,2,3,4 --alps=0.5,1,2"
 echo "systematics $runlabel --start_id=1001 --pdf=$pdfsets $scalevars" | ./bin/madevent
 popd
 
+# check lhe output  
+echo -e "\nRun xml check" 
+xmllint --stream --noout ${LHEWORKDIR}/process/madevent/Events/${runlabel}/events.lhe ; test $? -eq 0 || exit 1 
+echo "Number of weights that are NaN:" 
+grep  NaN  ${LHEWORKDIR}/process/madevent/Events/${runlabel}/events.lhe | grep "</wgt>" | wc -l ; test $? -eq 0 || exit 1 
+echo -e "All checks passed \n" 
+
+# copy output and print directory 
 mv ${LHEWORKDIR}/process/madevent/Events/${runlabel}/events.lhe ${LHEWORKDIR}/cmsgrid_final.lhe
-
 ls -l
-echo
 
+# exit 
 exit 0
